Explicit cast to (ssize_t) for stats p->ncopies
It's exteremely unlikely that we'll have more than SSIZE_MAX copies of a single
chunk, but I figured it was worth adding a check anyway.  If this is so
unlikely that it's a probabilistic impossibility, I'll happily remove the
check.

POSIX only guarantees that SSIZE_MAX is >= 32767; our code (elsewhere)
guarantees that SSIZE_MAX is >= MAXCHUNK (which is 261120).  On Debian 7 Wheezy
(released 2013, LTS until 2018), SSIZE_MAX is 2^31-1, so the #if define will be
useful on at least one platform.

This is the only place in our code that calls chunks_stats_add() with anything
other than +1 or -1 as the final (ssize_t) argument.